Generalization AISW using fixest fixed effects regressions

Augmented IPW, generalization, or transport estimator with fast fixest regressions and computation in `data.table`, no regularization, and no sample splitting. Imputes counterfactual outcome for each observation i under each treament a as \(Y^a = \frac{S}{\rho(X)} \frac{A = a}{\pi^a (X)} (Y - \hat{\mu}^a(X)) + \mu^a(X) \) Where the first term is 1 for all observations under no sample selection, and therefore this is the doubly-robust Augmented Inverse Propensity Weighting (AIPW) estimator. When S is supplied, the argument in 'target' is used to fit either the generalization or transportation estimator. Recommended with the nonparametric/bayesian bootstrap for inference.

ateGTreg(
  d,
  yn = "y",
  an = "a",
  sn = NULL,
  xn = "1",
  fe = "0",
  target = c("generalize", "transport", "insample")
)

Arguments

d

data.table

yn

outcome name

an

treatment indicator name

sn

selection indicator name (null by default - this fits the AIPW regression)

xn

list of covariates (default intercept)

fe

list of fixed effects (default 0)

target

estimand (generalization / transportation/ insample)

Value

generalization effect and influence functions