prepare X matrix with flexible interactions and polynomials

Prepare matrix with polynomial basis expansion and/or interactions.

polySieveM(
  dat,
  dummies = NULL,
  continuouses = NULL,
  corr_cut = 0.9,
  k = 2,
  m = 2,
  raw = F
)

Arguments

dat

data table / dataframe

dummies

Names of dummy vars (null by default - auto-detected)

continuouses

Names of continuous vars to modify fn form (null by default - auto detected)

corr_cut

cutoff for correlation threshold to drop one of the vars (default = 0.9)

k

order of interactions: defaults to pairwise interactions

m

order of functions: defaults to quadratic functions

raw

raw or orthogonal bases

Value

data.frame with base terms and interactions + basis